requirements.txt
# CapeAI Enterprise Platform - Production Dependencies
# Python 3.11+ Compatible - HEROKU TESTED VERSIONS

# Core FastAPI Stack (HEROKU TESTED)
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database & ORM (HEROKU TESTED)
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
greenlet==3.0.1
alembic==1.13.0

# Authentication & Security (HEROKU TESTED)
PyJWT==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Payment Processing - HEROKU COMPATIBLE
stripe==7.7.0

# Email Validation & Cryptography (HEROKU TESTED)
python-dotenv==1.0.0
email-validator==2.1.0
cryptography==42.0.8

# Redis for Caching & Session Management
redis==5.0.1

# AI Provider Dependencies (HEROKU TESTED)
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
tiktoken==0.5.2

# ML Dependencies - HEROKU TESTED
numpy==1.26.4
scikit-learn==1.4.0
scipy==1.12.0

# Enterprise Monitoring & Analytics (HEROKU TESTED)
psutil==5.9.6
structlog==23.2.0
rich==13.7.0

# AWS SDK for S3 access (HEROKU TESTED)
boto3==1.34.0

# Email & Notifications (HEROKU TESTED)
aiosmtplib==2.0.2
jinja2==3.1.2

# HTTP Client for API integrations (HEROKU TESTED)
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# File and Content Processing (HEROKU TESTED)
aiofiles==23.2.1
pillow==10.2.0
bleach==6.1.0
markdown==3.5.1
textstat==0.7.3

# Date and Time Handling (HEROKU TESTED)
python-dateutil==2.8.2
pytz==2023.3

# Core Python Dependencies (HEROKU TESTED)
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
dnspython==2.4.2
h11==0.14.0
idna==3.6
six==1.16.0
sniffio==1.3.0
typing-extensions==4.9.0

# Production Server & Deployment (HEROKU TESTED)
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk[fastapi]==1.38.0

# Rate Limiting and Security (HEROKU TESTED)
slowapi==0.1.9

# Additional Utilities (HEROKU TESTED)
colorama==0.4.6
backoff==2.2.1
cachetools==5.3.2
validators==0.22.0

# User Agent Parsing (HEROKU TESTED)
user-agents==2.2.0
